[change] Add initial personal sub pages
