+++ src/App.js
@@ -9,7 +9,12 @@ import ContactPage from './components/contactPage/ContactPage';
 import SearchPage from './components/searchPage/SearchPage';
 import LoginPage from './components/loginPage/LoginPage';
 import SignupPage from './components/signupPage/SingupPage';
-import PersonalPage from './components/personalPage/PersonalPage';
+import ListPage from './components/personalPage/ListPage';
+import BadgesPage from './components/personalPage/BadgesPage';
+import ReceivedPage from './components/personalPage/ReceivedPage';
+import ToReceivePage from './components/personalPage/toReceivePage';
+import CollectedSharesPage from './components/personalPage/CollectedSharesPage';
+import CollectedRestaurantsPage from './components/personalPage/CollectedRestaurantsPage';
 
 import './App.css';
 
@@ -39,8 +44,23 @@ const App = () => {
         <Route exact path="/signup">
           <SignupPage />
         </Route>
-        <Route exact path="/personal">
-          <PersonalPage />
+        <Route exact path="/personal/list">
+          <ListPage />
+        </Route>
+        <Route exact path="/personal/badges">
+          <BadgesPage />
+        </Route>
+        <Route exact path="/personal/received">
+          <ReceivedPage />
+        </Route>
+        <Route exact path="/personal/toReceive">
+          <ToReceivePage />
+        </Route>
+        <Route exact path="/personal/collectedShares">
+          <CollectedSharesPage />
+        </Route>
+        <Route exact path="/personal/collectedRestaurants">
+          <CollectedRestaurantsPage />
         </Route>
         <Route path="/">
           <LandingPage />
