+++ src/components/personalPage/Badges.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import styled from 'styled-components';
+
+const Badges = () => {
+  return <div>Badges</div>;
+};
+
+export default Badges;
