+++ src/components/personalPage/CollectedRestaurants.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import styled from 'styled-components';
+
+const CollectedRestaurants = () => {
+  return <div>CollectedRestaurants</div>;
+};
+
+export default CollectedRestaurants;
