+++ src/components/personalPage/CollectedShares.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import styled from 'styled-components';
+
+const CollectedShares = () => {
+  return <div>CollectedShares</div>;
+};
+
+export default CollectedShares;
