+++ src/components/personalPage/Received.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import styled from 'styled-components';
+
+const Received = () => {
+  return <div>Received</div>;
+};
+
+export default Received;
