+++ src/components/personalPage/toReceive.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import styled from 'styled-components';
+
+const ToReceive = () => {
+  return <div>ToReceive</div>;
+};
+
+export default ToReceive;
